Fixed bug in arbitrary transaction builder

+++ src/main/java/org/qortal/arbitrary/ArbitraryDataTransactionBuilder.java
@@ -112,7 +112,9 @@ public class ArbitraryDataTransactionBuilder {
             return transactionData;
 
         } catch (DataException e) {
-            arbitraryDataFile.deleteAll();
+            if (arbitraryDataFile != null) {
+                arbitraryDataFile.deleteAll();
+            }
             throw(e);
         }
